refactor: product/dto/display 내 DTO를 product 공용 DTO로 변경
